Added support for skin server to change Added the ability to revert side/edge textures to default (by sending -1) Created an Advanced Options section, for controlling new options Merged a seconds set of player hacks into the client Added an advanced option to switch between hack modes Fixed "/" opening a chat window where the caretPos was still 0 Added water level to advanced options Added sky color to advanced options Added cloud color to advanced options Added fog color to advanced options Added sunlight color to advanced options Added shadow color to advanced options

+++ src/main/java/com/mojang/minecraft/net/PacketType.java
@@ -119,10 +119,10 @@ public class PacketType {
 	EXT_ADD_ENTITY = new PacketType(new Class[] { Byte.TYPE, String.class,
 		String.class });
 	EXT_REMOVE_PLAYER_NAME = new PacketType(new Class[] { Short.TYPE });
-	ENV_SET_COLOR = new PacketType(new Class[] { Byte.TYPE, Integer.TYPE, Integer.TYPE, Integer.TYPE });
+	ENV_SET_COLOR = new PacketType(new Class[] { Byte.TYPE, Short.TYPE, Short.TYPE, Short.TYPE });
 	SELECTION_CUBOID = new PacketType(new Class[] { Byte.TYPE,
 		String.class, Short.TYPE, Short.TYPE, Short.TYPE, Short.TYPE,
-		Short.TYPE, Short.TYPE, Integer.TYPE, Integer.TYPE, Integer.TYPE, Integer.TYPE });
+		Short.TYPE, Short.TYPE, Short.TYPE, Short.TYPE, Short.TYPE, Short.TYPE });
 	REMOVE_SELECTION_CUBOID = new PacketType(new Class[] { Byte.TYPE });
 	SET_BLOCK_PERMISSIONS = new PacketType(new Class[] { Byte.TYPE, Byte.TYPE, Integer.TYPE });
 	CHANGE_MODEL = new PacketType(new Class[] { Byte.TYPE, String.class });
